+++ packages/sources/svelte/src/components/forms/VtmnTextInput/VtmnTextInput.svelte
@@ -1,6 +1,6 @@
 <script>
   import { cn } from '../../../utils/classnames';
-  import { VtmnIcon } from '../../..';
+  import VtmnIcon from '../../../guidelines/iconography/VtmnIcon/VtmnIcon.svelte';
 
   /** @restProps { input | textarea } */